Began adding test for PostGIS class
Do this to ensure test coverage of existing methods
