+++ geopandas_postgis/tests/resources/cb_2017_us_state_500k.cpg
@@ -1,0 +1,1 @@
+UTF-8
