+++ geopandas_postgis/tests/resources/cb_2017_us_state_500k.prj
@@ -1,0 +1,1 @@
+GEOGCS["GCS_North_American_1983",DATUM["D_North_American_1983",SPHEROID["GRS_1980",6378137,298.257222101]],PRIMEM["Greenwich",0],UNIT["Degree",0.017453292519943295]]
